requirements/dev.txt
# This file pulls in everything a developer needs. If it's a basic package
# needed to run the site, it belongs in requirements/prod.txt. If it's a
# package for developers (testing, docs, etc.), it goes in this file.
-r prod.txt

# Utils
ipython

# Deployment
Fabric
django-supervisor
uWSGI

# Documentation
Sphinx

# debug toolbar
django-debug-toolbar==1.0.1
django-debug-toolbar-template-timings==0.6

